Rename lib/richlist/richlist.js to lib/richlist/controller.js
