+++ lib/index.js
@@ -14,7 +14,7 @@ var UtilsController = require('./utils');
 var CurrencyController = require('./currency');
 var ZerocoinController = require('./zerocoin');
 var ExtendedController = require('./extended');
-var RichListController = require('./richlist/richlist');
+var RichListController = require('./richlist/controller');
 var RateLimiter = require('./ratelimiter');
 var morgan = require('morgan');
 var bitcore = require('zcore-lib');
