+++ lib/richlist/controller.js
