--- test/data/richlist/chain2.json
@@ -1,24 +1,0 @@
-[
-    {"height":0},
-    {
-        "hash": "b596c54581727d73071cc7f1aa1d56469f745e48d8456b190e189a4046ca5623",
-        "height": 1,
-        "txids": [
-           "6e3baad5e5b9d270e0c66981f633e3433c502826f3b89f2533f52131aab8a1f6"
-        ]
-    },
-    {
-        "hash": "2d10520434078bc57cf2cd2ef974c3aa448559ada4c033397d4d4ddcff9608be",
-        "height": 2,
-        "txids": [
-           "e48ac666d59401db247d509000cdb9c89fccccbe7d2cf408da42b79a01bc608e"
-        ]
-    },
-    {
-        "hash": "70b3eeb19aa77580224475e377c222c19481727ffb1994b5825691c95fd0c900",
-        "height": 3,
-        "txids": [
-            "b3c939b9a6f049e73efbbe916650054ecbe5651e1c2394b6e2084fedcee05729"
-        ]
-    }
-]
