--- test/data/richlist/transactions.json
@@ -1,83 +1,0 @@
-[
-    {
-        "hash": "e48ac666d59401db247d509000cdb9c89fccccbe7d2cf408da42b79a01bc608e",
-        "height": 2,
-        "inputs": [
-            {
-                "prevTxId": null,
-                "outputIndex": null,
-                "address": null,
-                "satoshis": null
-            }
-        ],
-        "inputSatoshis": 0,
-        "outputs": [
-            {
-                "satoshis": 4000000000,
-                "address": "aNUjTa4XLrCpRL5hqJf8Y4T6Cn3pZLLRUH"
-            },
-            {
-                "satoshis": 200000000,
-                "address": "a1kCCGddf5pMXSipLVD9hBG2MGGVNaJ15U"
-            }
-        ]
-    },
-    {
-        "hash": "6e3baad5e5b9d270e0c66981f633e3433c502826f3b89f2533f52131aab8a1f6",
-        "height": 1,
-        "inputs": [
-            {
-                "prevTxId": null,
-                "outputIndex": null,
-                "address": null,
-                "satoshis": null
-            }
-        ],
-        "outputs": [
-            {
-                "satoshis": 4000000000,
-                "address": "a9hZRxDCTomprkk4ajNUbGCGJbTTnXNcR5"
-            },
-            {
-                "satoshis": 200000000,
-                "address": "a1kCCGddf5pMXSipLVD9hBG2MGGVNaJ15U"
-            }
-        ]
-    },
-    {
-        "hash": "b3c939b9a6f049e73efbbe916650054ecbe5651e1c2394b6e2084fedcee05729",
-        "height": 3,
-        "inputs": [
-            {
-                "prevTxId": "e48ac666d59401db247d509000cdb9c89fccccbe7d2cf408da42b79a01bc608e",
-                "outputIndex": 0,
-                "address": "aNUjTa4XLrCpRL5hqJf8Y4T6Cn3pZLLRUH",
-                "satoshis": 4000000000
-            }
-        ],
-        "outputs": [
-            {
-                "satoshis": 4000000000,
-                "address": "aGm5jzFHLCt4pQLGwoHD2WijDmqbXfeWCz"
-            }
-        ]
-    },
-    {
-        "hash": "2b8bf3f3eab3a7b5efd524ac2bec7e7d11e8a1921ef4f26d1b4790e7d6e11bc2",
-        "height": 3,
-        "inputs": [
-            {
-                "prevTxId": "6e3baad5e5b9d270e0c66981f633e3433c502826f3b89f2533f52131aab8a1f6",
-                "outputIndex": 0,
-                "address": "a9hZRxDCTomprkk4ajNUbGCGJbTTnXNcR5",
-                "satoshis": 4000000000
-            }
-        ],
-        "outputs": [
-            {
-                "satoshis": 4000000000,
-                "address": "aGm5jzFHLCt4pQLGwoHD2WijDmqbXfeWCz"
-            }
-        ]
-    }
-]
